requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
xgboost
mlflow
dagshub
python-dotenv
jupyter
notebook
ipykernel
pytest
black
flake8
isort
openpyxl
joblib
ensure
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors